Consolidate commands/responses into commands.rs and responses.rs
These were previously split into separate files because each one
required its own special serialization logic.

Now that we're using serde we can combine them all into a single file
which, as of right now, is a little over 100 lines each. Not too shabby.
