+++ src/responses.rs
@@ -1,0 +1,118 @@
+//! Responses to commands sent from the HSM, intended as part of the public
+//! API of this crate.
+
+pub use failure::Error;
+pub(crate) use securechannel::CommandType;
+use serde::ser::Serialize;
+use serde::de::DeserializeOwned;
+
+use {Algorithm, Capabilities, Domains, ObjectId, ObjectLabel, ObjectOrigin, ObjectType, SequenceId};
+
+pub(crate) trait Response: Serialize + DeserializeOwned + Sized {
+    /// Command ID this response is for
+    const COMMAND_TYPE: CommandType;
+}
+
+/// Response from `CommandType::DeleteObject`
+///
+/// <https://developers.yubico.com/YubiHSM2/Commands/Delete_Object.html>
+#[derive(Serialize, Deserialize, Debug)]
+pub struct DeleteObjectResponse {}
+
+impl Response for DeleteObjectResponse {
+    const COMMAND_TYPE: CommandType = CommandType::DeleteObject;
+}
+
+/// Response from `CommandType::Echo`
+///
+/// <https://developers.yubico.com/YubiHSM2/Commands/Echo.html>
+#[derive(Serialize, Deserialize, Debug)]
+pub struct EchoResponse {
+    /// Echo response
+    pub message: Vec<u8>,
+}
+
+impl Response for EchoResponse {
+    const COMMAND_TYPE: CommandType = CommandType::Echo;
+}
+
+/// Response from `CommandType::GenAsymmetricKey`
+///
+/// <https://developers.yubico.com/YubiHSM2/Commands/Generate_Asymmetric_Key.html>
+#[derive(Serialize, Deserialize, Debug)]
+pub struct GenAsymmetricKeyResponse {
+    /// ID of the key
+    pub key_id: ObjectId,
+}
+
+impl Response for GenAsymmetricKeyResponse {
+    const COMMAND_TYPE: CommandType = CommandType::GenAsymmetricKey;
+}
+
+/// Response from `CommandType::GetObjectInfo`
+///
+/// <https://developers.yubico.com/YubiHSM2/Commands/Get_Object_Info.html>
+#[derive(Serialize, Deserialize, Debug)]
+pub struct GetObjectInfoResponse {
+    /// Capabilities
+    pub capabilities: Capabilities,
+
+    /// Object identifier
+    pub id: u16,
+
+    /// Length of object in bytes
+    pub length: u16,
+
+    /// Domains from which object is accessible
+    pub domains: Domains,
+
+    /// Object type
+    pub object_type: ObjectType,
+
+    /// Algorithm this object is intended to be used with
+    pub algorithm: Algorithm,
+
+    /// Sequence: number of times an object with this key ID and type has
+    /// previously existed
+    pub sequence: SequenceId,
+
+    /// How did this object originate? (generated, imported, etc)
+    pub origin: ObjectOrigin,
+
+    /// Label of object
+    pub label: ObjectLabel,
+
+    /// Delegated Capabilities
+    pub delegated_capabilities: Capabilities,
+}
+
+impl Response for GetObjectInfoResponse {
+    const COMMAND_TYPE: CommandType = CommandType::GetObjectInfo;
+}
+
+/// Response from `CommandType::ListObjects`
+///
+/// <https://developers.yubico.com/YubiHSM2/Commands/List_Objects.html>
+#[derive(Serialize, Deserialize, Debug)]
+pub struct ListObjectsResponse {
+    /// Objects in the response
+    pub objects: Vec<ListObjectsEntry>,
+}
+
+/// Brief information about an object as returned from the `ListObjects` command
+#[derive(Serialize, Deserialize, Debug)]
+pub struct ListObjectsEntry {
+    /// Object identifier
+    pub id: ObjectId,
+
+    /// Object type
+    pub object_type: ObjectType,
+
+    /// Sequence: number of times an object with this key ID and type has
+    /// previously existed
+    pub sequence: SequenceId,
+}
+
+impl Response for ListObjectsResponse {
+    const COMMAND_TYPE: CommandType = CommandType::ListObjects;
+}
