--- src/responses/delete_object.rs
@@ -1,13 +1,0 @@
-//! Response from `CommandType::DeleteObject`
-
-use super::{CommandType, Response};
-
-/// Response from `CommandType::DeleteObject`
-///
-/// <https://developers.yubico.com/YubiHSM2/Commands/Delete_Object.html>
-#[derive(Serialize, Deserialize, Debug)]
-pub struct DeleteObjectResponse {}
-
-impl Response for DeleteObjectResponse {
-    const COMMAND_TYPE: CommandType = CommandType::DeleteObject;
-}
