--- src/responses/gen_asymmetric_key.rs
@@ -1,17 +1,0 @@
-//! Response from `CommandType::GenAsymmetricKey`
-
-use ObjectId;
-use super::{CommandType, Response};
-
-/// Response from `CommandType::GenAsymmetricKey`
-///
-/// <https://developers.yubico.com/YubiHSM2/Commands/Generate_Asymmetric_Key.html>
-#[derive(Serialize, Deserialize, Debug)]
-pub struct GenAsymmetricKeyResponse {
-    /// ID of the key
-    pub key_id: ObjectId,
-}
-
-impl Response for GenAsymmetricKeyResponse {
-    const COMMAND_TYPE: CommandType = CommandType::GenAsymmetricKey;
-}
