--- src/responses/list_objects.rs
@@ -1,31 +1,0 @@
-//! Response from `CommandType::ListObjects`
-
-use {ObjectId, ObjectType, SequenceId};
-use super::{CommandType, Response};
-
-/// Response from `CommandType::ListObjects`
-///
-/// <https://developers.yubico.com/YubiHSM2/Commands/List_Objects.html>
-#[derive(Serialize, Deserialize, Debug)]
-pub struct ListObjectsResponse {
-    /// Objects in the response
-    pub objects: Vec<ListObjectsEntry>,
-}
-
-/// Brief information about an object as returned from the `ListObjects` command
-#[derive(Serialize, Deserialize, Debug)]
-pub struct ListObjectsEntry {
-    /// Object identifier
-    pub id: ObjectId,
-
-    /// Object type
-    pub object_type: ObjectType,
-
-    /// Sequence: number of times an object with this key ID and type has
-    /// previously existed
-    pub sequence: SequenceId,
-}
-
-impl Response for ListObjectsResponse {
-    const COMMAND_TYPE: CommandType = CommandType::ListObjects;
-}
